refactor(interpreter): split activation rollback tests
