+++ src/interpreter/src/activation/tests/dispatch.rs
@@ -1,4 +1,27 @@
-use super::*;
+use super::support::{
+    FLAT_TUPLE_ACTIVATION,
+    MechTuple,
+    NESTED_TUPLE_ACTIVATION,
+    REPEATED_CAPTURE_ACTIVATION,
+    Ref,
+    Value,
+    ValueKind,
+    assert_dispatch_turn,
+    hash_str,
+    interpret,
+    interpret_more,
+    load_atom_tuple_activation,
+    load_enum_activation,
+    plan_snapshot,
+    root_cell,
+    set_atom_tuple_event,
+    set_enum_event,
+    set_f64_matrix_event,
+    set_tuple_event,
+    set_unit_enum_event,
+    tuple_fixture,
+    turn_executed_nodes,
+};
 
     #[test]
     fn activation_pattern_selects_pressed_released_and_wildcard() {
