+++ src/interpreter/src/activation/tests/exhaustiveness.rs
@@ -1,4 +1,22 @@
-use super::*;
+use super::support::{
+    ActivationPatternArmsNonExhaustive,
+    ActivationPatternWildcardMustBeLast,
+    Ref,
+    Value,
+    arm_register_nodes,
+    assert_dispatch_turn,
+    committed_capture_value,
+    f64_symbol,
+    interpret,
+    interpret_more,
+    plan_snapshot,
+    registration,
+    root_cell,
+    selected_arm_index,
+    set_f64_matrix_event,
+    set_f64_symbol,
+    set_tuple_event,
+};
 
     #[test]
     fn activation_final_binding_is_exhaustive_and_guarded_binding_falls_through_to_it() {
