+++ src/interpreter/src/activation/tests/registers.rs
@@ -1,4 +1,18 @@
-use super::*;
+use super::support::{
+    ReactiveDependencyKind,
+    Ref,
+    Value,
+    arm_pulse_generation,
+    arm_register_nodes,
+    committed_capture_value,
+    f64_symbol,
+    interpret,
+    plan_snapshot,
+    registration,
+    root_cell,
+    selected_arm_index,
+    set_f64_symbol,
+};
 
     #[test]
     fn activation_selected_arm_commits_register_batch_and_other_arms_schedule_nothing() {
